% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cost_HQ.R
\name{cost_HQ}
\alias{cost_HQ}
\title{The cost of HQ optimization}
\usage{
cost_HQ(H, Q, beta, lambda)
}
\arguments{
\item{H}{The matrix in the optimization.}

\item{Q}{The vector in the optimization.}

\item{beta}{The variable in the optimization.}

\item{lambda}{The penalty factor in the optimization.}
}
\value{
\item{cost}{The cost value.}
}
\description{
The cost function of HQ optimization.
}
