% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gra_HQ.R
\name{gra_HQ}
\alias{gra_HQ}
\title{The gradient of HQ optimization}
\usage{
gra_HQ(H, Q, beta)
}
\arguments{
\item{H}{The matrix in the optimization.}

\item{Q}{The vector in the optimization.}

\item{beta}{The variable in the optimization.}
}
\value{
\item{gra}{The gradient value.}
}
\description{
The gradient of HQ optimization.
}
