% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ista_HQ.R
\name{ista_HQ}
\alias{ista_HQ}
\title{Solve HQ optimization with ista algorithm}
\usage{
ista_HQ(H, Q, lambda, beta_start = NA, conv_thresh = 1e-08)
}
\arguments{
\item{H}{The matrix in the optimization.}

\item{Q}{The vector in the optimization.}

\item{lambda}{The penalty factor in the optimization.}

\item{beta_start}{The initial value for the variable.}

\item{conv_thresh}{The relative error.}
}
\value{
\item {beta}{The solution for the HQ optimization.}
\item {costS}{The cost valueS at each step.}
}
\description{
Solve HQ optimization with ista algorithm.
}
