% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ls_penalty.R
\name{ls_penalty}
\alias{ls_penalty}
\title{Solve the penalty optimization for linear model}
\usage{
ls_penalty(X, y, penalty.type = "AdapLASSO", gamma = 1)
}
\arguments{
\item{X}{The covariate in the model.}

\item{y}{The response variable in the model.}

\item{penalty.type}{The penalty type, LASSO or Adaptive LASSO. The default value is AdapLASSO.}

\item{gamma}{The factor in the Adaptive LASSO. The default value is 1.}
}
\value{
\item{beta}{The solution of optimization with penalty.}
}
\description{
Solve the LASSO or Adaptive LASSO for linear model.
}
