% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onlineTL_ls.R
\name{onlineTL_ls}
\alias{onlineTL_ls}
\title{The updating for streaming batch data with TL}
\usage{
onlineTL_ls(
  XsXs,
  Xsys,
  ysys,
  ns,
  beta_s,
  Xt,
  yt,
  penalty.type = "AdapLASSO",
  gamma = 1
)
}
\arguments{
\item{XsXs}{The statistic \eqn{X^{T}X}.}

\item{Xsys}{The statistic \eqn{X^{T}y}.}

\item{ysys}{The statistic \eqn{y^{T}y}.}

\item{ns}{The number of source data.}

\item{beta_s}{The estimate from the source data.}

\item{Xt}{The covariavte from the target batch.}

\item{yt}{The response from the target batch.}

\item{penalty.type}{The penalty type, LASSO or Adaptive LASSO. The default value is AdapLASSO.}

\item{gamma}{The factor in the Adaptive LASSO. The default value is 1.}
}
\value{
\item{beta}{The solution of updating.}
\item{XsXs}{The updated statistic \eqn{X^{T}X}.}
\item{Xsys}{The updated statistic \eqn{X^{T}y}.}
\item{ysys}{The updated statistic \eqn{y^{T}y}.}
\item{ns}{The updated number of source data.}
}
\description{
The updating for streaming batch data with TL.
}
