% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thresh_fct.R
\name{thresh_fct}
\alias{thresh_fct}
\title{Thresholding function of LASSO}
\usage{
thresh_fct(z, lambda)
}
\arguments{
\item{z}{The constant in the loss .}

\item{lambda}{The penalty parameter.}
}
\value{
\item{fz}{The solution.}
}
\description{
Thresholding function of LASSO.
}
